pyproject.toml
[tool.poetry]
name = "quizbot"
version = "0.1.0"
description = ""
authors = ["Wiktor Flis <wikflis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
gradio = "^4.41.0"
pdfminer = "^20191125"
transformers = "^4.44.0"
pinecone-client = "^5.0.1"
langchain-pinecone = "^0.1.3"
langchain-community = "^0.2.11"
langchain-openai = "^0.1.21"
langchain = "^0.2.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
